+++ CluckADuck/Overlays/GameOverOverlay.cpp
@@ -82,7 +82,7 @@ void GameOverOverlay::onUpdate(float dt)
 
 void GameOverOverlay::onDraw(sf::RenderTarget& rt)
 {
-	sf::RectangleShape rect(sf::Vector2f(rt.getSize().x, rt.getSize().y));
+	sf::RectangleShape rect(sf::Vector2f(static_cast<float>(rt.getSize().x), static_cast<float>(rt.getSize().y)));
 	rect.setFillColor(sf::Color(0, 0, 0, 180));
 	rt.draw(rect);
