+++ CluckADuck/Overlays/PauseOverlay.hpp
@@ -32,7 +32,7 @@ public:
 	virtual void onUpdate(float dt) {}
 	virtual void onDraw(sf::RenderTarget& rt)
 	{
-		sf::RectangleShape rect(sf::Vector2f(rt.getSize().x, rt.getSize().y));
+		sf::RectangleShape rect(sf::Vector2f(static_cast<float>(rt.getSize().x), static_cast<float>(rt.getSize().y)));
 		rect.setFillColor(sf::Color(0, 0, 0, 180));
 		rt.draw(rect);
